+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Current (main branch)
 
 * `lvl app create`: if no organisation is given, your current organisation is used as default.
+* Basic support for app component attachments.
 
 ## 1.7.2
 * Fixed `lvl app delete` not accepting `-y` to confirm deletion.